requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
google-generativeai==0.4.0
reportlab==4.2.0
markdown==3.6
